+++ demisto_sdk/commands/common/errors.py
@@ -188,7 +188,8 @@ ERROR_CODE = {
     "pack_metadata_non_approved_usecases": {'code': "PA119", 'ui_applicable': False, 'related_field': ''},
     "pack_metadata_non_approved_tags": {'code': "PA120", 'ui_applicable': False, 'related_field': ''},
     "pack_metadata_price_change": {'code': "PA121", 'ui_applicable': False, 'related_field': ''},
-    "is_wrong_usage_of_usecase_tag": {'code': "PA122", 'ui_applicable': False, 'related_field': ''},
+    "pack_name_already_exists": {'code': "PA122", 'ui_applicable': False, 'related_field': ''},
+    "is_wrong_usage_of_usecase_tag": {'code': "PA123", 'ui_applicable': False, 'related_field': ''},
     "readme_error": {'code': "RM100", 'ui_applicable': False, 'related_field': ''},
     "image_path_error": {'code': "RM101", 'ui_applicable': False, 'related_field': ''},
     "readme_missing_output_context": {'code': "RM102", 'ui_applicable': False, 'related_field': ''},
@@ -1185,8 +1186,14 @@ class Errors:
 
     @staticmethod
     @error_code_decorator
+    def pack_name_already_exists(new_pack_name) -> str:
+        return f"A pack named: {new_pack_name} already exists in content repository, " \
+               f"change the pack's name in the metadata file."
+
+    @staticmethod
+    @error_code_decorator
     def is_wrong_usage_of_usecase_tag():
-        return "pack_metadata.json file contains the Use Case tag, without having any PB, Acidents Types or Layouts"
+        return "pack_metadata.json file contains the Use Case tag, without having any PB, incidents Types or Layouts"
 
     @staticmethod
     @error_code_decorator