+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-#Changelog
+# Changelog
 * Added FeedHelloWorld template option for *--template* flag in **demisto-sdk init** command.